requirements.txt
annotated-types>=0.7.0
anyio>=4.10.0
blinker>=1.9.0
cachetools>=5.5.2
certifi>=2025.8.3
cffi>=1.17.1
charset-normalizer>=3.4.3
click==8.1.8
Flask>=3.1.2
google-ai-generativelanguage>=0.6.18
google-api-core>=2.25.1
google-api-python-client>=2.179.0
google-auth>=2.40.3
google-auth-httplib2>=0.2.0
google-cloud-speech>=2.33.0
google-cloud-texttospeech>=2.27.0
google-genai>=1.31.0
googleapis-common-protos>=1.70.0
grpcio>=1.74.0
grpcio-status>=1.74.0
h11>=0.16.0
httpcore>=1.0.9
httplib2>=0.22.0
httpx>=0.28.1
idna>=3.10
itsdangerous>=2.2.0
Jinja2>=3.1.6
MarkupSafe>=3.0.2
numpy==2.0.2
opencv-python>=4.12.0.88
pillow>=11.3.0
pocketsphinx>=5.0.4
proto-plus>=1.26.1
protobuf>=6.32.0
pvporcupine>=3.0.5
pyasn1>=0.6.1
pyasn1_modules>=0.4.2
PyAudio>=0.2.14
pycparser>=2.22
pydantic>=2.11.7
pydantic_core>=2.33.2
pydub>=0.25.1
pyparsing>=3.2.3
requests>=2.32.5
rsa>=4.9.1
setuptools>=80.9.0
sniffio>=1.3.1
sounddevice>=0.5.2
tenacity>=9.1.2
tqdm>=4.67.1
typing-inspection>=0.4.1
typing_extensions>=4.14.1
uritemplate>=4.2.0
urllib3>=2.5.0
websockets>=15.0.1
Werkzeug>=3.1.3
